+++ fv3gfs/__init__.py
@@ -6,26 +6,30 @@ from ._wrapper import (
     compute_physics, apply_physics,
 )
 from ._restart import get_restart_names, open_restart
-from ._ghost_cells import without_ghost_cells, with_ghost_cells
 from fv3util import (
     InvalidQuantityError, DYNAMICS_PROPERTIES, PHYSICS_PROPERTIES,
     read_state, write_state, apply_nudging, get_nudging_tendencies, ZarrMonitor,
-    CubedSpherePartitioner
+    CubedSpherePartitioner, TilePartitioner, CubedSphereCommunicator, TileCommunicator,
+    Communicator,
+    Quantity, X_DIMS, Y_DIMS, HORIZONTAL_DIMS, INTERFACE_DIMS, X_DIM, X_INTERFACE_DIM,
+    Y_DIM, Y_INTERFACE_DIM, Z_DIM, Z_INTERFACE_DIM, UnitsError
 )
 
 
 __all__ = [
     'initialize', 'step', 'step_dynamics', 'step_physics', 'save_intermediate_restart_if_enabled',
+    'compute_physics', 'apply_physics',
     'save_fortran_restart',
     'cleanup', 'get_state', 'set_state',
     'get_n_ghost_cells', 'get_step_count', 'get_tracer_metadata',
-    'without_ghost_cells', 'with_ghost_cells',
     'InvalidQuantityError',
     'DYNAMICS_PROPERTIES', 'PHYSICS_PROPERTIES',
     'read_state', 'write_state', 'get_restart_names',
     'apply_nudging', 'get_nudging_tendencies', 'open_restart', 'ZarrMonitor',
-    'CubedSpherePartitioner',
-    'compute_physics', 'apply_physics',
+    'CubedSpherePartitioner', 'TilePartitioner', 'CubedSphereCommunicator', 'TileCommunicator',
+    'Communicator',
+    'Quantity', 'X_DIMS', 'Y_DIMS', 'HORIZONTAL_DIMS', 'INTERFACE_DIMS', 'X_DIM', 'X_INTERFACE_DIM',
+    'Y_DIM', 'Y_INTERFACE_DIM', 'Z_DIM', 'Z_INTERFACE_DIM', 'UnitsError'
 ]
 
 __version__ = '0.3.1'